requirements.txt
# FastAPI and core dependencies
annotated-types==0.7.0
anyio==3.7.1
click==8.3.1
fastapi==0.104.1
h11==0.16.0
httptools==0.7.1
idna==3.11
pydantic==2.12.4
pydantic_core==2.41.5
pydantic-settings==2.1.0
python-dotenv==1.2.1
PyYAML==6.0.3
sniffio==1.3.1
starlette==0.27.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1

# File upload and processing
python-multipart==0.0.6
pandas>=2.2.0
openpyxl==3.1.2
aiofiles==23.2.1

# Database
sqlalchemy==2.0.23
asyncmy==0.2.9
greenlet>=3.0.0

# Dash frontend
dash==2.14.2
dash-bootstrap-components==1.5.0
plotly==5.18.0
requests==2.31.0
